feat: security refactor, UI improvements, and project setup

+++ src/main/java/com/main/bitebyte/security/JwtConfig.java
@@ -10,13 +10,13 @@ public class JwtConfig {
     private String jwtSecret;
 
     @Value("${jwt.expiration}")
-    private int jwtExpirationMs;
+    private long jwtExpirationMs;
 
     public String getJwtSecret() {
         return jwtSecret;
     }
 
-    public int getJwtExpirationMs() {
+    public long getJwtExpirationMs() {
         return jwtExpirationMs;
     }
 }
